refactor: move to `services/users/notify_mentioned_users`
