+++ app/controllers/site_notices_controller.rb
@@ -23,7 +23,7 @@ class SiteNoticesController < ApplicationController
 
     if @site_notice.save
       @site_notice.create_activity :create, current_user
-      NotifyUsersMentionedIn[@site_notice.message, attached: @site_notice, notifier: current_user]
+      Users::NotifyMentionedUsers[@site_notice.message, attached: @site_notice, notifier: current_user]
       redirect_to @site_notice, notice: "Successfully created site notice."
     else
       render :new
@@ -36,7 +36,7 @@ class SiteNoticesController < ApplicationController
   def update
     if @site_notice.update(site_notice_params)
       @site_notice.create_activity :update, current_user
-      NotifyUsersMentionedIn[@site_notice.message, attached: @site_notice, notifier: current_user]
+      Users::NotifyMentionedUsers[@site_notice.message, attached: @site_notice, notifier: current_user]
       redirect_to @site_notice, notice: "Successfully updated site notice."
     else
       render :edit
